+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.2.4] - 2020-01-05  
+
+#### Added
+- A new property for the rules: "Top frame only" (enabled by default). 
+  - if *enabled* the rule will be injected ONLY in the top frame of the page (when the selector matches the tab url, as it has done up to now)
+  - if *disabled* the rule will be injected in iframes too.
+
+#### Changed
+- Replaced the old "Active Tab data" system to handle async calls and navigations in the background.js
 
 ## [0.2.3] - 2019-04-28  
 
@@ -73,8 +82,8 @@ The initial Beta release
 
 
 
-[0.2.3]: https://github.com/Lor-Saba/Code-Injector/releases/tag/v0.2.1
-[0.2.2]: https://github.com/Lor-Saba/Code-Injector/releases/tag/v0.2.1
+[0.2.3]: https://github.com/Lor-Saba/Code-Injector/releases/tag/v0.2.3
+[0.2.2]: https://github.com/Lor-Saba/Code-Injector/releases/tag/v0.2.2
 [0.2.1]: https://github.com/Lor-Saba/Code-Injector/releases/tag/v0.2.1
 [0.2.0]: https://github.com/Lor-Saba/Code-Injector/releases/tag/v0.2.0
 [0.1.1]: https://github.com/Lor-Saba/Code-Injector/releases/tag/v0.1.1